+++ cherrypy/__main__.py
@@ -1,5 +1,5 @@
+"""CherryPy'd cherryd daemon runner."""
 from cherrypy.daemon import run
 
 
-if __name__ == '__main__':
-    run()
+__name__ == '__main__' and run()