+++ cherrypy/_cptree.py
@@ -7,11 +7,10 @@ import six
 import cherrypy
 from cherrypy._cpcompat import ntou
 from cherrypy import _cpconfig, _cplogging, _cprequest, _cpwsgi, tools
-from cherrypy.lib import httputil
+from cherrypy.lib import httputil, reprconf
 
 
 class Application(object):
-
     """A CherryPy Application.
 
     Servers and gateways should not instantiate Request objects directly.
@@ -32,7 +31,7 @@ class Application(object):
     """A dict of {path: pathconf} pairs, where 'pathconf' is itself a dict
     of {key: value} pairs."""
 
-    namespaces = _cpconfig.NamespaceSet()
+    namespaces = reprconf.NamespaceSet()
     toolboxes = {'tools': cherrypy.tools}
 
     log = None
@@ -47,6 +46,7 @@ class Application(object):
     relative_urls = False
 
     def __init__(self, root, script_name='', config=None):
+        """Initialize Application with given root."""
         self.log = _cplogging.LogManager(id(self), cherrypy.log.logger_root)
         self.root = root
         self.script_name = script_name
@@ -61,6 +61,7 @@ class Application(object):
             self.merge(config)
 
     def __repr__(self):
+        """Generate a representation of the Application instance."""
         return '%s.%s(%r, %r)' % (self.__module__, self.__class__.__name__,
                                   self.root, self.script_name)
 
@@ -144,17 +145,17 @@ class Application(object):
 
         try:
             req.close()
-        except:
+        except Exception:
             cherrypy.log(traceback=True, severity=40)
 
         cherrypy.serving.clear()
 
     def __call__(self, environ, start_response):
+        """Call a WSGI-callable."""
         return self.wsgiapp(environ, start_response)
 
 
 class Tree(object):
-
     """A registry of CherryPy applications, mounted at diverse points.
 
     An instance of this class may also be used as a WSGI callable
@@ -170,6 +171,7 @@ class Tree(object):
     WSGI callable if you happen to be using a WSGI server)."""
 
     def __init__(self):
+        """Initialize registry Tree."""
         self.apps = {}
 
     def mount(self, root, script_name='', config=None):
@@ -216,8 +218,17 @@ class Tree(object):
             app = Application(root, script_name)
 
             # If mounted at "", add favicon.ico
-            if script_name == '' and root is not None and not hasattr(root, 'favicon_ico'):
-                favicon = os.path.join(os.getcwd(), os.path.dirname(__file__), 'favicon.ico')
+            needs_favicon = (
+                script_name == ''
+                and root is not None
+                and not hasattr(root, 'favicon_ico')
+            )
+            if needs_favicon:
+                favicon = os.path.join(
+                    os.getcwd(),
+                    os.path.dirname(__file__),
+                    'favicon.ico',
+                )
                 root.favicon_ico = tools.staticfile.handler(favicon)
 
         if config:
@@ -257,6 +268,7 @@ class Tree(object):
             path = path[:path.rfind('/')]
 
     def __call__(self, environ, start_response):
+        """Pre-initialize WSGI env and call WSGI-callable."""
         # If you're calling this, then you're probably setting SCRIPT_NAME
         # to '' (some WSGI servers always set SCRIPT_NAME to '').
         # Try to look up the app using the full path.