+++ cherrypy/lib/httpauth.py
@@ -24,9 +24,10 @@ SUPPORTED_QOP - list of supported 'Digest' 'qop'.
 import time
 from hashlib import md5
 
+from six.moves.urllib.request import parse_http_list, parse_keqv_list
+
 from cherrypy._cpcompat import (
     base64_decode, ntob,
-    parse_http_list, parse_keqv_list
 )
 
 
@@ -278,7 +279,9 @@ def _computeDigestResponse(auth_map, password, method='GET', A1=None,
     algorithm = params.get('algorithm', MD5)
 
     H = DIGEST_AUTH_ENCODERS[algorithm]
-    KD = lambda secret, data: H(secret + ':' + data)
+
+    def KD(secret, data):
+        return H(secret + ':' + data)
 
     qop = params.get('qop', None)
 
@@ -339,7 +342,8 @@ def _checkBasicResponse(auth_map, password, method='GET', encrypt=None,
                         **kwargs):
     # Note that the Basic response doesn't provide the realm value so we cannot
     # test it
-    pass_through = lambda password, username=None: password
+    def pass_through(password, username=None):
+        return password
     encrypt = encrypt or pass_through
     try:
         candidate = encrypt(auth_map['password'], auth_map['username'])