+++ cherrypy/lib/static.py
@@ -1,12 +1,12 @@
+"""Module with helpers for serving static files."""
+
 import os
 import re
 import stat
 import mimetypes
 
-try:
-    from io import UnsupportedOperation
-except ImportError:
-    UnsupportedOperation = object()
+from email.generator import _make_boundary as make_boundary
+from io import UnsupportedOperation
 
 from six.moves import urllib
 
@@ -35,7 +35,6 @@ def serve_file(path, content_type=None, disposition=None, name=None,
     to the basename of path. If disposition is None, no Content-Disposition
     header will be written.
     """
-
     response = cherrypy.serving.response
 
     # If path is relative, users should fix it by making path absolute.
@@ -117,7 +116,6 @@ def serve_fileobj(fileobj, content_type=None, disposition=None, name=None,
     serve_fileobj(), expecting that the data would be served starting from that
     position.
     """
-
     response = cherrypy.serving.response
 
     try:
@@ -190,12 +188,6 @@ def _serve_fileobj(fileobj, content_type, content_length, debug=False):
             else:
                 # Return a multipart/byteranges response.
                 response.status = '206 Partial Content'
-                try:
-                    # Python 3
-                    from email.generator import _make_boundary as make_boundary
-                except ImportError:
-                    # Python 2
-                    from mimetools import choose_boundary as make_boundary
                 boundary = make_boundary()
                 ct = 'multipart/byteranges; boundary=%s' % boundary
                 response.headers['Content-Type'] = ct
@@ -321,6 +313,13 @@ def staticdir(section, dir, root='', match='', content_types=None, index='',
     section = section.rstrip(r'\/')
     branch = request.path_info[len(section) + 1:]
     branch = urllib.parse.unquote(branch.lstrip(r'\/'))
+
+    # On Windows requesting a file in sub-dir of the staticdir results
+    # in mixing of delimiter styles, eg: C:\static\js/script.js
+    # Python normally converts this but not when the staticdir is
+    # supplied in long-path notation, eg: \\?\C:\static\js/script.js
+    if os.name == 'nt':
+        branch = branch.replace('/', '\\')
 
     # If branch is "", filename will end in a slash
     filename = os.path.join(dir, branch)